+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## ChangeLog
 
+* Rename `SetLogFile` to `SetFile` to reduce redundancy.
+
 * `Transaction` now implements an `Ignore` method which will prevent
   any of the transaction's data from being recorded.